+++ app/routers/register.py
@@ -1,0 +1,36 @@
+from fastapi import status, Depends, HTTPException, APIRouter
+from sqlalchemy.orm import Session
+from logging import getLogger
+
+from app.schemas import UserCreateReq
+from app.utils import hash_password, check_token
+import app.models as models
+from app.database import get_db
+from app.rabbitmq import publish_message
+
+router = APIRouter()
+logger = getLogger("uvicorn")
+
+
+@router.post("/register", status_code=status.HTTP_201_CREATED)
+def create_user(req: UserCreateReq, db: Session = Depends(get_db)) -> None:
+    user_data = req.user_data
+    token_data = req.token_data
+    exception = check_token(token=token_data.token, email=user_data.email,
+                            req_action=token_data.action, db=db)
+    if exception:
+        raise exception
+
+    user_exists = db.query(models.User).filter(
+        models.User.email == user_data.email).first()
+    if user_exists:
+        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST,
+                            detail=f"メールアドレス{user_data.email} "
+                                   f"はすでに登録されています。")
+    user_data.password = hash_password(user_data.password)
+    new_user = models.User(**user_data.dict())
+    db.add(new_user)
+    db.commit()
+    db.refresh(new_user)
+    publish_message(message={"email": user_data.email},
+                    queue_name="confirm_register_email")
